create a queryValidator
