+++ apps/backend/src/controller/project/projectsController.ts
@@ -1,6 +1,6 @@
-import { ErrorJSON, Monitor, ProjectService } from '@tee/backend-ddd'
+import { ErrorJSON, Monitor, ProjectService, NAF1Letters } from '@tee/backend-ddd'
 import { Controller, Get, Queries, Res, Route, SuccessResponse, Tags, TsoaResponse } from 'tsoa'
-import { ProjectFilterQuery } from '@tee/common'
+import { ProjectFilterQuery, Sector } from '@tee/common'
 import { Project } from '@tee/data'
 
 @SuccessResponse('200', 'OK')
@@ -13,8 +13,12 @@ export class ProjectsController extends Controller {
   @Get()
   public async get(
     @Queries() filterData: ProjectFilterQuery,
-    @Res() requestFailedResponse: TsoaResponse<500, ErrorJSON>
+    @Res() requestFailedResponse: TsoaResponse<400 | 500, ErrorJSON>
   ): Promise<Project[]> {
+    if (!this._validateData(filterData)) {
+      return requestFailedResponse(400, { message: 'Invalid filter data provided' })
+    }
+
     const projectResults = new ProjectService().getFiltered(filterData)
 
     if (projectResults.isErr) {
@@ -26,5 +30,18 @@ export class ProjectsController extends Controller {
     }
 
     return projectResults.value
+  }
+
+  private _validateData(filterData: ProjectFilterQuery): boolean {
+    if (filterData.codeNAF1 && !NAF1Letters.includes(filterData.codeNAF1)) {
+      return false
+    }
+
+    const validSectors = Object.values(Sector)
+    if (filterData.sector && !validSectors.includes(filterData.sector)) {
+      return false
+    }
+
+    return true
   }
 }
