+++ libs/backend-ddd/src/common/index.ts
@@ -2,3 +2,4 @@ export * from './controller/jsonError'
 export { default as ServiceNotFoundError } from './domain/api/serviceNotFoundError'
 export { ensureError } from './domain/error/errors'
 export { default as Monitor } from './domain/monitoring/monitor'
+export * from './naf1'
