+++ libs/backend-ddd/src/program/infrastructure/types.ts
@@ -1,4 +1,4 @@
-import { PublicodesKeys, Sector, PublicodeObjective } from '@tee/common'
+import { PublicodesKeys, PublicodeObjective } from '@tee/common'
 import { PublicodesCondition } from '@tee/data'
 export interface PublicodesInputData {
   [PublicodesKeys.CodeNAF]?: string
@@ -17,17 +17,6 @@ export enum PublicodesQuestionnaireRoute {
   NoSpecificGoal = '"je ne sais pas par où commencer"',
   SpecificGoal = '"objectif précis"'
 }
-
-export const SectorToNAFSection: Record<Sector, string[]> = {
-  [Sector.Craftsmanship]: ['C', 'F', 'G'],
-  [Sector.Industry]: ['B', 'C', 'D', 'E'],
-  [Sector.Tourism]: ['I'],
-  [Sector.Tertiary]: ['G', 'H', 'J', 'K', 'L', 'M', 'N', 'O', 'P', 'Q', 'R', 'S', 'T', 'U'],
-  [Sector.Agriculture]: ['A'],
-  [Sector.Other]: ['D', 'E', 'F', 'H', 'J', 'K', 'L', 'M', 'N', 'O', 'P', 'Q', 'R', 'S', 'T', 'U']
-}
-
-export const NAF1Letters = [...'ABCDEFGHIJKLMNOPQRSTU'] as const
 
 export type ObjectivePublicodeData = {
   [key in PublicodesCondition]: PublicodeObjective[]
